Bug#25930 (CREATE TABLE x SELECT ... parses columns wrong when ran with            ANSI_QUOTES mode)
This patch contains a test case only, showing that the bug has been fixed.

The issue was related to parsing <"blah">, where the lexical analyser
would not properly delimit the start and end of the token.

Properly making the token boundaries has been fixed in sql_lex.cc
with the patch for bug 25411 : see the Lex_input_stream class.