Bug#30747 Create table with identical constraint names behaves incorrectly
MySQL provides what appears to be a non standard extension to the
FOREIGN KEY syntax which let users name (label/tag) a foreign key
to more easily identify a specific foreign key if any problems show
up later during the query parsing or execution. But the foreign key
name was not being properly set to the right key, possible leaving
the foreign key with no name.