snprintf() should always return non-negative result. According to Microsoft documentation about _vscprintf():
  If format is a null pointer, the invalid parameter handler is invoked,
  as described in Parameter Validation. If execution is allowed to
  continue, the functions return -1 and set errno to EINVAL.

The UNIX variant of snprintf() segfaults if format is a NULL pointer
(similar to strlen(NULL) for example), so it is better to conform to
this behavior and crash our custom Windows version instead of
returning -1. Noone would expect -1 to be returned from snprintf().

Cosmetic: Add a space after typecast.

Approved by:	Marko