branches/zip: innodb-index.result: Adapt the results to a patch from MySQL. Unfortunately, the patch seems to prevent MySQL from utilizing secondary indexes in the ORDER BY clauses.  Thus, the test no longer checks if newly created indexes are well-formed.
ChangeSet@1.2528.98.1  2007-08-02 12:45:56-07:00  igor@mysql.com

Fixed bug#28404.
This patch adds cost estimation for the queries with ORDER BY / GROUP BY
and LIMIT. 
If there was a ref/range access to the table whose rows were required
to be ordered in the result set the optimizer always employed this access
though a scan by a different index that was compatible with the required 
order could be cheaper to produce the first L rows of the result set.
Now for such queries the optimizer makes a choice between the cheapest
ref/range accesses not compatible with the given order and index scans
compatible with it.