Bug #30695: An apostrophe ' in the comment of the ADD PARTITION   causes the Server to crash.
Accessing partitioned table with an apostrophe in partition options
like DATA DIRECTORY, INDEX DIRECTORY or COMMENT causes server crash.

Partition options were saved in .frm file without escaping.
When accessing such table it is not possible to properly restore
partition information.
Crashed because there was no check for partition info parser failure.

Fixed by escaping quoted text in the partition info when writing it to
the frm-file and added a check that it was able to parse the partition
info before using it 

NOTE: If the comment is written by an earlier version of the server,
the corrupted frm-file is not fixed, but left corrupted, you have to
manually drop the table and recreate it.


mysql-test/r/partition.result:
  bug #30695: An apostrophe ' in the comment of the ADD PARTITION causes
  the Server to crash.
   testresult
mysql-test/t/partition.test:
  bug #30695: An apostrophe ' in the comment of the ADD PARTITION causes
  the Server to crash.
    testcase
sql/sql_partition.cc:
  Bug #30695: An apostrophe ' in the comment of the ADD PARTITION causes
    the Server to crash.
  
    Crashes when there is an non escaped apostrophe in the partition options
  
    fixed by escaping quoted text before writing to the frm-file
sql/table.cc:
  Bug #30695 An apostrophe ' in the comment of the ADD PARTITION
    causes the Server to crash
  
  problem was using a null poiter without check -> crash.
  
  added a check that the previus call succeded